Azure_functions/collecte_csv/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas 
pyarrow 
requests 
python-dotenv 
azure-storage-file-datalake